Windows,JNI: check cmd length for CreateProcess
Error out if the command we try to pass to
CreateProcess is longer than the limit.

Doing so results in a nicer error message than
"The parameter is incorrect" which is confusing.

In this commit I also improve the error reporting
of CreateProcessWithExplicitHandles.

See https://github.com/bazelbuild/bazel/issues/4083
See https://github.com/bazelbuild/bazel/issues/4096

Change-Id: I00ec52238706fd8140483eddb488c3069eaa7814
PiperOrigin-RevId: 175969789